Re-factoring to allow adding more datasets, improve state management, add prettier config, placeholders for precipitation
