+++ .prettierrc
@@ -1,6 +1,7 @@
 {
-    "trailingComma": "none",
-    "tabWidth": 4,
     "semi": true,
-    "singleQuote": true
-}
+    "trailingComma": "none",
+    "singleQuote": true,
+    "jsxSingleQuote": true,
+    "printWidth": 80
+  }
