+++ src/components/TourButton.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const StyledButton = styled.button`
+  position: absolute;
+  height: 32px;
+  width: 32px;
+  top: 170px;
+  right: 15px;
+  background-color: #242424;
+  color: #adadad;
+  border: none;
+  cursor: pointer;
+  z-index: 50;
+`;
+
+export default function TourButton({ onClick }) {
+  return <StyledButton onClick={onClick}>?</StyledButton>;
+}
